perf(parser): fuse the code-span newline probe into the closer scan (-3% parse)
A code span's content rules turn line endings into spaces, so
`normalize_code_span` asked `raw.contains('\n')` — a second full scan of
bytes the closer search had just walked. Almost every inline code span is
single-line, so almost every one paid a whole extra scan to prove it had
nothing to normalize. Ablating that one call was worth 1.3-2.9% of a parse.

The closer search now uses `memchr2(b'`', b'\n')` and carries the answer
out with it. Newlines inside a span are rare, so the extra iterations cost
nothing measurable, and `memchr2` is no slower than `memchr` per call.

Parse, best of 60 iterations:

  rust-book            2.699 -> 2.615 ms  -3.1%
  typescript-handbook  3.688 -> 3.629 ms  -1.6%
  vite-docs            1.216 -> 1.201 ms  -1.2%
  vue-docs             2.131 -> 2.108 ms  -1.1%

Parse + render:

  typescript-handbook  5.679 -> 5.601 ms  -1.4%
  vite-docs            1.742 -> 1.722 ms  -1.1%
  vue-docs             3.206 -> 3.187 ms  -0.6%
  rust-book            3.860 -> 3.849 ms  -0.3%

`parse_inline_code` and `normalize_code_span` move to an
`inline/code_span.rs` submodule to keep `inline.rs` inside the 350-line
source cap (362 -> 288).
